Correções importantes e implementações importantes.

+++ oms/ServiceDefaults/PositionState.cs
@@ -6,7 +6,7 @@ namespace ServiceDefaults;
 [StructLayout(LayoutKind.Explicit, Size = 16)]
 public struct PositionState
 {
-    [FieldOffset(0)] public int TotalQuantity;    // Custódia total disponível
-    [FieldOffset(4)] public int BlockedQuantity;  // Quantidade retida em ordens abertas de venda
-    [FieldOffset(8)] public int TradedQuantity;   // Execuções do dia
+    [FieldOffset(0)] public long TotalQuantity;    // Custódia total disponível
+    [FieldOffset(4)] public long BlockedQuantity;  // Quantidade retida em ordens abertas de venda
+    [FieldOffset(8)] public long TradedQuantity;   // Execuções do dia
 }
